ahoy_backend/requirements.txt
black==21.7b0
channels==3.0.4
channels_redis==3.3.0
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
django-stubs==1.9.0
prospector==1.5.1
pytest==6.2.4
rope==0.19.0
selenium==3.141.0
psycopg2-binary==2.9.1
pytest-django==4.4.0
pytest-env==0.6.2
mypy==0.910
types-selenium==3.141.3
